requirements.txt
setuptools
python-dotenv
KekikTaban

spotipy

Pyrogram
tgcrypto
asyncio
aiohttp

APScheduler